create rlhf model
